+++ core/src/main/java/org/svip/compare/conflicts/ConflictFactory.java
@@ -1,5 +1,8 @@
 package org.svip.compare.conflicts;
 
+import org.svip.metrics.resultfactory.Text;
+import org.svip.metrics.resultfactory.enumerations.INFO;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -10,6 +13,7 @@ import java.util.Set;
  * Handles creating new Conflicts
  *
  * @author Derek Garcia
+ * @author Thomas Roman
  */
 public class ConflictFactory {
     private final List<Conflict> conflicts = new ArrayList<>();
@@ -70,7 +74,7 @@ public class ConflictFactory {
             }
             // targetValue not in other set
             if (!compared)
-                addConflict(field, MismatchType.MISSING, "Contains " + field + " Data", null);
+                addConflict(field, MismatchType.MISSING, targetValue.toString(), null);
         }
 
         // Round 2: Don't compare other against target, just checking if present
@@ -88,7 +92,7 @@ public class ConflictFactory {
 
             // otherValue not in target set
             if (!compared)
-                addConflict(field, MismatchType.MISSING, null, "Contains " + field + " Data");
+                addConflict(field, MismatchType.MISSING, null, otherValue.toString());
         }
     }
 
@@ -130,22 +134,25 @@ public class ConflictFactory {
         if (!comparable(field, target, other))
             return;
 
+        // Construct Text to use for diff report conflict messages
+        Text text = new Text("Conflict", field);
+
         // Round 1: Compare target against other if equal
         for (String targetAlg : target.keySet()) {
             // If other doesn't contain hash, add as missing
             if (!other.containsKey(targetAlg)) {
-                addConflict(field, MismatchType.MISSING, "Contains " + field + " Data", null);
+                addConflict(field, MismatchType.MISSING, targetAlg, null);
                 continue;
             }
             // Compare hash values
-            addConflict(field + " " + targetAlg + " Hash", MismatchType.HASH_MISMATCH, target.get(targetAlg), other.get(targetAlg));
+            addConflict(text.getMessage(INFO.DIFF_HASH_ALG, targetAlg), MismatchType.HASH_MISMATCH, target.get(targetAlg), other.get(targetAlg));
         }
 
         // Round 2: Don't compare other against target, just checking if present
         for (String otherAlg : other.keySet()) {
             // If target doesn't contain hash, add as missing
             if (!target.containsKey(otherAlg))
-                addConflict(field, MismatchType.MISSING, null, "Contains " + field + " Data");
+                addConflict(field, MismatchType.MISSING, null, otherAlg);
         }
     }
 
@@ -167,7 +174,8 @@ public class ConflictFactory {
         // One is missing from the other
         // TODO Better way to handle this case
         if (target == null || other == null) {
-            addConflict(field, MismatchType.MISSING, (target == null ? "Present" : null), (other == null ? "Present" : null));
+            addConflict(field, MismatchType.MISSING, (target == null ? other.toString() : null),
+                                                     (other == null ? target.toString() : null));
             return false;
         }