+++ core/src/main/java/org/svip/metrics/resultfactory/enumerations/INFO.java
@@ -5,6 +5,7 @@ package org.svip.metrics.resultfactory.enumerations;
  * Enumeration that holds the types of information a result could have
  *
  * @author Matthew Morrison
+ * @author Thomas Roman
  */
 public enum INFO {
     MISSING,
@@ -14,5 +15,6 @@ public enum INFO {
     NULL,
     ERROR,
     MATCHING,
-    NOT_MATCHING
+    NOT_MATCHING,
+    DIFF_HASH_ALG
 }