+++ core/src/test/java/org/svip/sbom/model/objects/SVIPComponentObjectConflictsTest.java
@@ -1,6 +1,8 @@
 package org.svip.sbom.model.objects;
 
 import org.junit.jupiter.api.Test;
+import org.svip.metrics.resultfactory.Text;
+import org.svip.metrics.resultfactory.enumerations.INFO;
 import org.svip.sbom.builder.objects.SVIPComponentBuilder;
 import org.svip.sbom.factory.objects.SVIPSBOMComponentFactory;
 import org.svip.sbom.model.interfaces.generics.Component;
@@ -154,9 +156,11 @@ public class SVIPComponentObjectConflictsTest {
         for(Conflict c : conflictList)
         {
             if(c.GetType() == MismatchType.MISSING && Objects.equals(c.GetMessage(), "Component Hash is missing"))
-                if(Objects.equals(c.GetTarget(),"Contains Component Hash Data") && c.GetOther() == null)
+                if(Objects.equals(c.GetTarget(), "SHA1") &&
+                        c.GetOther() == null)
                     c1 = true;
-                else if(c.GetTarget() == null && Objects.equals(c.GetOther(), "Contains Component Hash Data"))
+                else if(c.GetTarget() == null &&
+                        Objects.equals(c.GetOther(), "SHA2"))
                     c2 = true;
         }